feat: Add shadcn utility function
This commit adds a new utility function called `cn` to the `shadcn.ts` file in the `src/app/utils` directory. The `cn` function is responsible for merging and applying Tailwind CSS classes using the `clsx` and `tailwind-merge` libraries. This utility function will be used to simplify class merging in the codebase.